added commands to fetch approved licenses and  alpine licenses and build a json from it to be embedded by component service
