+++ cmd/devguard-cli/main.go
@@ -55,6 +55,7 @@ func init() {
 	rootCmd.AddCommand(commands.NewVulndbCommand())
 	rootCmd.AddCommand(commands.NewComponentsCommand())
 	rootCmd.AddCommand(commands.NewDaemonCommand())
+	rootCmd.AddCommand(commands.NewLicensesCommand())
 }
 
 func main() {
